requirements.txt
pymongo
python-dotenv
gunicorn
flask
flask-cors
cloudinary
pillow